[common|core|station|node]: provide provisioning IPs
There is now a new ip address space in plantmesh.

core: 10.11.0.1
station|node: 10.11.0.2

core and station assign their IPs to their eth0 (WAN) interface. node
uses an alias on the existing bridge.

This is in preparation for a provisioning script on the core which will
be launched on eth0 link up and automatically try to provision
10.11.0.2.

We really should create a provisioning vlan in the future.

Also fix a related bug setting the hostname while eth0 is down.